backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
langgraph==0.2.36
langchain-core==0.3.33
langsmith==0.2.11
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
tenacity==8.5.0
huggingface_hub==0.26.5
